feat: add install mods page component
